+++ kong/client.go
@@ -35,10 +35,12 @@ var defaultCtx = context.Background()
 // Kong cluster
 type Client struct {
 	client                  *http.Client
-	defaultRootURL          string
+	baseRootURL             string
 	workspace               string       // Do not access directly. Use Workspace()/SetWorkspace().
 	workspaceLock           sync.RWMutex // Synchronizes access to workspace.
 	common                  service
+	ConsumerGroupConsumers  AbstractConsumerGroupConsumerService
+	ConsumerGroups          AbstractConsumerGroupService
 	Consumers               AbstractConsumerService
 	Developers              AbstractDeveloperService
 	DeveloperRoles          AbstractDeveloperRoleService
@@ -57,6 +59,9 @@ type Client struct {
 	RBACRoles               AbstractRBACRoleService
 	RBACEndpointPermissions AbstractRBACEndpointPermissionService
 	RBACEntityPermissions   AbstractRBACEntityPermissionService
+	Vaults                  AbstractVaultService
+	Keys                    AbstractKeyService
+	KeySets                 AbstractKeySetService
 
 	credentials       abstractCredentialService
 	KeyAuths          AbstractKeyAuthService
@@ -123,9 +128,11 @@ func NewClient(baseURL *string, client *http.Client) (*Client, error) {
 	if err != nil {
 		return nil, fmt.Errorf("parsing URL: %w", err)
 	}
-	kong.defaultRootURL = url.String()
+	kong.baseRootURL = url.String()
 
 	kong.common.client = kong
+	kong.ConsumerGroupConsumers = (*ConsumerGroupConsumerService)(&kong.common)
+	kong.ConsumerGroups = (*ConsumerGroupService)(&kong.common)
 	kong.Consumers = (*ConsumerService)(&kong.common)
 	kong.Developers = (*DeveloperService)(&kong.common)
 	kong.DeveloperRoles = (*DeveloperRoleService)(&kong.common)
@@ -144,6 +151,9 @@ func NewClient(baseURL *string, client *http.Client) (*Client, error) {
 	kong.RBACRoles = (*RBACRoleService)(&kong.common)
 	kong.RBACEndpointPermissions = (*RBACEndpointPermissionService)(&kong.common)
 	kong.RBACEntityPermissions = (*RBACEntityPermissionService)(&kong.common)
+	kong.Vaults = (*VaultService)(&kong.common)
+	kong.Keys = (*KeyService)(&kong.common)
+	kong.KeySets = (*KeySetService)(&kong.common)
 
 	kong.credentials = (*credentialService)(&kong.common)
 	kong.KeyAuths = (*KeyAuthService)(&kong.common)
@@ -191,9 +201,9 @@ func (c *Client) Workspace() string {
 // baseURL build the base URL from the rootURL and the workspace
 func (c *Client) workspacedBaseURL(workspace string) string {
 	if len(workspace) > 0 {
-		return c.defaultRootURL + "/" + workspace
+		return c.baseRootURL + "/" + workspace
 	}
-	return c.defaultRootURL
+	return c.baseRootURL
 }
 
 // DoRAW executes an HTTP request and returns an http.Response
@@ -226,7 +236,8 @@ func (c *Client) DoRAW(ctx context.Context, req *http.Request) (*http.Response, 
 func (c *Client) Do(ctx context.Context, req *http.Request,
 	v interface{},
 ) (*Response, error) {
-	resp, err := c.DoRAW(ctx, req)
+	// TODO https://github.com/Kong/go-kong/issues/273 clear this lint ignore
+	resp, err := c.DoRAW(ctx, req) //nolint:bodyclose
 	if err != nil {
 		return nil, err
 	}
@@ -360,4 +371,58 @@ func (c *Client) RootJSON(ctx context.Context) ([]byte, error) {
 	}
 
 	return body, nil
+}
+
+func (c *Client) BaseRootURL() string {
+	return c.baseRootURL
+}
+
+// ReloadDeclarativeRawConfig sends out the specified config to configured Admin
+// API endpoint using the provided reader which should contain the JSON
+// serialized body that adheres to the configuration format specified at:
+// https://docs.konghq.com/gateway/latest/production/deployment-topologies/db-less-and-declarative-config/#declarative-configuration-format
+// It returns the response body and an error, if it encounters any.
+func (c *Client) ReloadDeclarativeRawConfig(
+	ctx context.Context,
+	config io.Reader,
+	checkHash bool,
+	flattenErrors bool,
+) ([]byte, error) {
+	type sendConfigParams struct {
+		CheckHash     int `url:"check_hash,omitempty"`
+		FlattenErrors int `url:"flatten_errors,omitempty"`
+	}
+	var checkHashI int
+	if checkHash {
+		checkHashI = 1
+	}
+	var flattenErrorsI int
+	if flattenErrors {
+		flattenErrorsI = 1
+	}
+	req, err := c.NewRequest(
+		"POST",
+		"/config",
+		sendConfigParams{CheckHash: checkHashI, FlattenErrors: flattenErrorsI},
+		config,
+	)
+	if err != nil {
+		return nil, fmt.Errorf("creating new HTTP request for /config: %w", err)
+	}
+
+	resp, err := c.DoRAW(ctx, req)
+	if err != nil {
+		return nil, fmt.Errorf("failed posting new config to /config: %w", err)
+	}
+	defer resp.Body.Close()
+
+	b, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, fmt.Errorf("could not read /config %d status response body: %w", resp.StatusCode, err)
+	}
+	if resp.StatusCode < 200 || resp.StatusCode >= 400 {
+		return b, fmt.Errorf("failed posting new config to /config: got status code %d and body %s", resp.StatusCode, b)
+	}
+
+	return b, nil
 }